JAT-0008 Rename helpers.py to helpers_jira.py
